Begin fixing landing customization
There is a surprising amount of content that depends on `$.Site.Data.landing`.
This is bad and should be removed or made to use something default-but-changeable
This changeset does exactly that. I've also reworked the way that iamge compare
is checked for, which makes it possible to use it in the body content as well
later as a shortcode.

This also begins the process of making all scripts fingerprintable for SRI.